Move component PositiveNumberByDevelopedDateCard from 'components/cards' to 'components/index/CardsReference'
